Added some basic player movement(WIP)
Added a vector 2 class which can be altered easier than SDL_Fpoint. Using this class the player now has a velocity and direction with the idea being that when the up arrow is pressed velocity is added in the direction the player is facing. However it does not currently work as the direction vector is not right.