Fixed player movement
Added a Lerp function to fix the player 'snapping' to a direction when moving at max_speed. Instead the x and y velocity of the player moves from their current velocity toward the target velocity by a certain amount each frame. This gives the game a more floaty feeling where the player can drift slightly. Also added friction to the player so they slow down when not moving.